0413 SO Blob형식으로 생성 외 1
- SO blob형식으로 생성
- Save, Load함수 파일명 다르게 생성/로드 가능하게 변경
- HeavilyModified 코드 로드 되게끔 변경

+++ Assets/Scripts/Runtime/Item/UsableDetailData.cs
@@ -10,11 +10,11 @@ using UnityEngine;
 public class UsableDetailData : ItemDetailData
 {
     [SerializeField] public List<short> durationList;
-    [SerializeField] public List<byte> powerList;
+    [SerializeField] public List<short> powerList;
     [SerializeField] public List<ChargeInfo> chargeInfoList = new List<ChargeInfo>();
 
     public short Duration(byte index) => durationList[index];
-    public byte Power(byte index) => powerList[index];
+    public short Power(byte index) => powerList[index];
     public ChargeInfo ChargeInfo(byte index) => chargeInfoList[index];
 }
 
@@ -22,7 +22,7 @@ public struct UsableDetailBlobData
 {
     public byte index;
     public short duration;
-    public sbyte power;
+    public short power;
     public ChargeInfo chargeInfo;
 }
